Garbage-collect entry-memoized values
Currently, if a value isn't set to `@persist` or `@memoize`, it's still
memoized (at the TaskRunnerEntry level) for the duration of the `get`
call. This isn't a big deal because it's pretty rare to disable both
persistence and memoization for an entity. However, once we start using
tuple descriptors, those tuples will have both persistence and
memoization disabled *and* may contain large values that we don't want
to keep in memory.

In particular, for any entity defined like this:

    @builder
    @bn.memoize(False)
    @bn.outputs('a', 'b')
    def _(...):
        ...

while `a` and `b` will be correctly persisted and not memoized, the
tuple `a, b` *will* be memoized for the duration of the `get` call. If
`a` or `b` is large, this will be a problem for users.

To address this, I've added logic to detect when these entry-memoized
values are no longer needed and delete them. This should allow us to
enable tuple descriptors without any memory-usage regressions, and also
gives us some memory savings in the rare case where both persistence and
memoization are manually disabled.

This is implemented by allowing requirements to be discarded when
they're no longer needed. The criteria for discarding a requirement
depend on which of three types it falls into, as documented on the
EntryRequirements class. This system of state and requirements is
getting a little rickety again. I'm open to ideas for simplifying it; or
maybe it'll be easier after we've done of the future refactoring