Add AIP based execution
This change adds the capability in Bionic to use AIP based execution.
Unlike parallel execution, not all tasks are computed in parallel.
Instead, you specify which entities to compute on AIP using the
`@aip_process` decorator.

This change builds on top of #249 and uses the `aip` package from that
change. Since it is still a PoC, we haven't documented this behavior or
added it to the release notes.