Fix assisted versioning for non-persisted values
Looks like we had a bug where assisted versioning would not catch
versioning errors on non-persisted entities. This adds a fix by allowing
bytecode data to propagate through `Provenance.exact_hash`. This change
actually simplifies the matching model (instead of having four match
levels, we have three and they all propagate the same way), but it makes
it more challenging to report the source of the error when we find it,
since it could be somewhere upstream of where it's detected. To fix
*this*, each Provenance now has references to its dependency digests,
and each digest now has a reference to its provenance if it was created
from one, allowing us to search backward for the provenance with the
problem.

This also requires making ProvenanceDigests serializable as dicts, which
is getting a bit gross. I'm wondering if we can use a library like
`cattrs` to replace our bespoke object<->dict<->YAML serialization.