Clear old files when updating cache test vectors
Previously, the `generate_test_compatibility_cache` would create new
cache files but not delete the old ones, so that step had to be done
manually (or the files could just be left to accumulate in git). With
this change, the script deletes the old files before creating the new
ones.