Support versioning modes in random tests
This extends the new random test system to support the `assist` and
`auto` versioning modes, bringing us much closer to being able to get
rid of the old fuzz tests. (These new tests also would have caught the
failure of assisted versioning to work on non-persisted values,
presumably.)

This required the following supporting changes:
- CodeVersioningError now has a `bad_descriptor` field which identifies
the function that needs to be fixed.
- We now define each function in the flow under test by generating
Python code (`output_value_fragment`) instead of just calling out to a
separate function; this makes changes in the definition visible in the
Python bytecode.
- In the model flow, persisted values can be marked "stale", which will
cause them to trigger a CodeVersioningError.