Tweak `@changes_per_run` memoization check
This modifies how we check that all `@changes_per_run` tasks are
memoized. This will be important later, when we introduce draft
descriptors, because these tasks will be split into multiple sub-tasks
and the `@changes_per_run` decorator will no longer be on the same task
as the memoization setting.