Tweak task serializability check
This updates the way we check if a task can be executed remotely.
Currently we check if the task has the `is_simple_lookup` flag, which is
a proxy for whether the task can be cloudpickled and sent to a different
process. This changes introduces an explicit `can_be_serialized`
property, and we now check all the task's non-persistable ancestors for
this property as well.

Right now this makes no difference because all non-cloudpicklable tasks
(that is, the fixed-value lookup tasks) are persistable, but as we
split out tasks into smaller sub-tasks this will no longer be true.