Compute cache flags before TaskState creation
This change computes the persistence and memoization flags *before*
creating the task graph rather than after. This will allow us to
have the task structure depend on these settings; in particular, I want
to descriptors and tasks specifically for cached files, but that means I
need to know whether an entity is persisted before I create its tasks.

To do this, instead of fetching or generating an EntityDefinition for
each task, we instead create a DescriptorMetadata object that contains
the finalized caching flags. (This name also resolves the awkwardness of
having an EntityDefinition for things that aren't entities.)

I also switched the order of the `_prevalidate_base_dnodes` and
`_set_up_bootstrap` steps, because our metadata objects now depend
whether they're computed before or after the bootstrap; we need to make
sure to do the bootstrap setup before we construct all the other nodes.

There should be no functional changes.