Add new node type for artifact descriptors
This introduces a new type of descriptor: an artifact descriptor. This
descriptor corresponds to the Artifact class, and represents an entity
value that has been serialized and persisted to disk.

Artifact descriptors are implemented as a special case of a new type of
descriptor node: a GenericNode. These are written like this:

    some_entity/artifact

Here `artifact` is "generic", meaning it's parameterized by another
descriptor: there's a different artifact for each entity.
 Currently `artifact` is the only kind of generic, but in the future I
expect us to have other generics like `provenance`, `path`, etc.

Artifact descriptors won't be exposed to users of Bionic; they're just
an internal concept that allows us to decompose tasks into smaller
parts.

This commit only introduces the new classes and parsing logic for
artifact descriptors; it doesn't add or change any runnable code in
Bionic.