Remove Query class
I've removed the Query class: everything that was previously using the
class now uses just the subset of components (TaskKey, Provenance,
Protocol) that it needs. In many cases, just the Provenance is
sufficient.

The primary motivation here is that I want to pull the logic for
serializing/deserializing values to/from disk out of the PersistentCache
class in order to have separate file descriptor nodes, at which point
the cache won't care about protocols and it won't make sense for them to
be part of its API. More generally, it was always a little weird for to
be passing a protocol around as part of a query, since it's only needed
in one place.

A secondary advantage is that this frees up the term "query" to be used
for something else -- maybe we can use "query" to describe one call to
`flow.get`? (As in, "this value is cached for the duration of the
query".)