Add `baseline` marker to tests
This new marker identifies the "baseline" set of tests that run by
default. It's useful to be able to filter these out when we want to run
tests in multiple disjoint batches. For example:

    pytest
    pytest --parallel --slow -m 'not baseline'

The above will exclude all the baseline tests from the second batch,
making sure we don't run any test twice.