Introduce Artifact class
We often pass around a URL and a file content hash together; this commit
groups them together into one class.

Aside from simplifying our code, my goal is to make it easier to pull
the serialization/deserialization code out of PersistentCache. To do
that we need to stop passing around Results and only use the fields that
matter to the cache, which are now just the Provenance and the Artifact.

I've named the hash field `content_hash` even though elsewhere we call
it `value_hash`. My thinking here is that we're moving to a world where
the persistent cache just deals in files and doesn't know about
in-memory values, so it makes more sense to talk about "content" than
"value".