Update cache compatibility test vector
This doesn't change any of the exact hashes, but it does change two
things:

1. The JSON artifacts in the test vector no longer has newlines at the
end. These newlines were added by a rogue git hook, which means the
stored value hashes for these artifacts didn't actually match their
contents in git. We still need a way to systematically remove this git
hook for other contributors (it's installed Square-wide, and pre-commit
may be doing the same thing), but for now this fixes the issue, which
will become a problem in the next commit.
2. The metadata files are updated to reflect the more current DAG
structure; but again, no provenance hashes have changed.