Fix AIP dispatch
This fixes a bug introduced by my earlier draft descriptors change.
When determining whether to send a task to AIP or not, we need to
consider not just whether that task has an AIP config, but also anything
in its remote subgraph has one.

I don't have the real AIP tests working on my machine, but hopefully we can
get the fake AIP tests to the point where they can catch this bug.