Pull saving/loading code out of persistent cache
Now all serialization and deserialization happens outside the
PersistentCache class, which no longer deals with Results or Protocols,
just Artifacts and Provenances. This will make it possible to have
separate tasks for loading and saving data.

Unfortunately there are still a couple of awkward cases where we need to
call into the PersistentCache to handle some details. We could refactor
this further, but I think the path forward will become clearer as we
pull more stuff out of the cache. (For example, maybe we'll move to
separate local and cloud caches.)