Use full filename to infer protocol
This changes how we detect which protocol to use for deserializing
values: instead of passing the file extension to the protocol, we now
let the protocol figure it out by examining the filename itself.

The problem with the previous approach is that identifying the "file
extension" is non-trivial because extensions can contain periods (like
"dask.pq"). The cache can only figure this out because it created the
file itself and knows the entire filename format; but if we want to
support reading from user-provided files, we won't be able to do this.
Instead, we just let the protocol figure it out: the protocol either
expects a particular format or knows a range of supported file file
extension which it can check in order.