Run GCS tests using real GCS in CI
This change uses `setup-gcloud` GitHub Action to set up gcloud on CI
and runs the GCS tests using both fake and actual GCS.

This was a lot easier to use than I expected. You just need to add the
service account key to secrets and use it to authenticate. For more
information on how to use it, please check the
[readme](https://github.com/GoogleCloudPlatform/github-actions/blob/master/setup-gcloud/README.md).

I also verified that this works by checking that the CI build doesn't
skip GCS tests. I can also see the cached files inside the new bionic
bucket in our team's project when the CI is running.