Avoid redundant GCS downloads
One of my earlier cache refactors seems to have introduced a bug where
we download artifacts from GCS more times than necessary. In particular,
we were downloading artifacts to disk even when a cloud artifact would
have sufficed. This fix updates the cache accessor API to distinguish
between requesting a local artifact and requesting any kind of artifact.

Also, all loading operations now also ensure that the returned artifact
is replicated at all cache layers. Previously the caller was responsible
for this, but it would be a little brittle to have the cache implicitly
create a local artifact and depend on the caller to remember to actually
save it.