Rename FlowState to FlowConfig
I'd like to free up the term "state" for the kind of stuff that gets
maintained over the lifetime of a flow instance: i.e., the things we
cache in EntityDeriver and the TaskStates. "FlowConfig" is not a perfect
name, but it's correct in the sense that it contains the "configuration"
of a Flow.