Rename Bootstrap to ExecutionCore
I don't love either name, but I'd like to reserve the name "bootstrap"
for the initial setup phase, rather than the object it produces. This
will make it possible to introduce a "bootstrap core" that has default
values for all the fields, instead of allowing the core to be None.