Add context.temp_result_cache
This moves our temporary caching into a central object that lives in the
ExecutionContext. The immediate advantage of this is that the temporary
cache is now visible to TaskState, which will make it easier to refactor
our compute logic. (In particular, I'd like to move the `compute()`
method onto TaskState, and have TaskStates be able to call compute() on
their dependency states, but this requires TaskStates to have access all
cached values.)

In the future, it might make sense to move the memoized TaskState.result
into another cache object as well, but that raises some more questions
and doesn't block any immediate work, so I've just put it in a TODO for
now.