Add generic eviction code
This generalizes our "vacate" code to make it safe to use on any
temporarily memoized value, not just drafts. Now temporarily memoized
values can be evicted from memory and recomputed later if necessarily.

(This commit doesn't change any behavior; it just makes it possible to
introduce changes later.)

Motivation
----------

There are two motivations for this:

1. I want to split out the persistent saving and loading operations into
separate tasks, so that instead of a persistent entity being computed
like this:

    <x>           -> x
    (draft value) -> (final value)

it will be computed like this:

    <x>           -> x/artifact -> x
    (draft value) -> (artifact) -> (final value)

The catch is, if `x` is not memoized, we want it to be loaded from disk
each time it's needed; that is, we want the loading function `x/artifact
-> x` to be recomputed each time. That means we need a way to evict `x`
from memory as soon as it gets used (but not any sooner!).

2. Recomputing uncached values each time they're used (described
[here](https://github.com/square/bionic/issues/296)).

Implementation
--------------

Both of these goals require us to be able to: compute an entry value;
hold it in memory long enough for a dependency to access it; delete it;
and then recompute it again if necessary. (Also, if the value is a
tuple, we need to make sure to keep it around just long enough for all
of its followup tasks to run. If it weren't for this extra constraint,
we could solve this by just computing these uncached values (and their
uncached dependencies) lazily each time; but that would mean any task
that produces a pair of values would always get run twice.)

Previously the only values we evicted from memory were draft values.
This was easy to do because we know that draft values are only ever
accessed by their followup tasks, so once the followups are done, it's
safe to delete the draft value forever. However, now we want
to delete values that might be accessed later, so we need to be able to
recompute them on demand. This required the following changes:

1. We need a mechanism to delete requirements that are no longer needed.
We delete a requirement if (a) the requirement's source entry is fully
completed (`CACHED`), or (b) the requirement is marked "transient" and
its destination is fully completed. ("Transient" requirements are for
things like followup tasks, where we need an entry to be completed but
there's no other entry waiting for it.)

2. We now have one extra priority level: some followup tasks are not
just `HIGH` but `TOP` priority. The idea here is that if you have a
tuple, it's more urgent to break it into smaller pieces than it is to
process those individual pieces. This approach helps us get those pieces
out of memory as early as possible.

3. We need to actually delete each temporarily-memoized value from
memory at the right time. This can be either right after a child task
uses the value, or right after computing a value that no one is waiting
for (which can happen with followup tasks). Note that we don't do this
eviction for actual entity values, because that would be a breaking
change; we still just do this for draft values. We use a new
`should_memoize_for_query` field on `DescriptorMetadata` to determine
which values we should do this for.