Check exact call counts in random tests
Previously our random tests only verified that we called the correct set
of user functions, but not that they were called the correct number of
times. Now we check this, except for certain cases where it's hard to
predict the right number of calls: during parallel execution, and when
fixing versioning errors.