Support disabled memoization in random tests
This extends our random tests to include both memoized and non-memoized
entities. (Previously all entities were memoized.) I also updated our
test loop to call `get` a random number of times after each update.
(Previously we called `get` exactly once per update, which meant we
never used the same flow instance twice, so we wouldn't be fully
exercising the in-memory cache.)

This also introduces a ModelExecutionContext class which maintains the
state of a `get` call. This makes it easier to manage our multiple
levels of caching, and also opens the door to moving the compute logic
out of the harness class altogether.