Use same node color for members of tuple
This commit updates our DAG visualization to group together nodes
generated by a common function (i.e., a tuple-generating function). This
means the nodes will be visually aligned and have the same color. I
think this clarifies the relationship between these nodes.

(We already do this for nodes that have the same descriptor but
different case keys; this just extends the grouping logic to nodes that
share entity names.)