Cache object hashes within CodeHasher
This change keeps track of all the object hashes for a CodeHasher
instance and avoids hashing the same object multiple times using it.
We scope the object hashes to a CodeHasher instance instead of globally
to avoid using a stale hash in case an object change by the time we
hash it again.