Remove TODO to hash module version
There is no sure way to get version for kinds of module in Python. You
can find versions for top level modules from `__version__` member
filed, but that version field doesn't exist for all modules. Since we
can't hash module versions consistently, this change does not hash the
version to be consistent and to avoid confusing behavior for the user.