Skip code object references for hashing
We find references for a function from it's code object using the code
context created from the function. These references won't necessarily
share the same context with the function.

This change stops sending the code context to hash the references. But
without a code context, we cannot hash a code object efficiently (since
we can't find it's references). For any references that are code
objects, this change also treats them as a complex type and emits a
warning instead.