Hash all constant types and more
For Smart Caching, we want to hash all kind of constants that exists in
`func.__code__.co_consts` variable. We already hash most of them, but
with this change, we should be hashing all of the constants. I couldn't
find an official source that lists the type of constants, but I looked
at CPython impl to see the different constant kinds. Particularly, I
looked at the function that validates constant in ast which should list
all constant types. But I also found another function that creates a key
for constants (doesn't have all the constants though).

validate_constant in AST: https://github.com/python/cpython/blob/master/Python/ast.c#L120
_PyCode_ConstantKey: https://github.com/python/cpython/blob/86e322f1412be6c5a6c8fa8e73af68b99c42bb18/Objects/codeobject.c#L784

This change also hashes a couple of more types that are not constants
but can be easily hashed.