Always emit versioning warnings when appropriate
It looks like we were defaulting to emitting versioning warnings for
functions decorated with `@version`, but not for undecorated functions.
This commit should make us always default to emitting warnings (unless
auto/assist versioning is not enabled).

To do this, I changed the `@version` decorator to defer to the defaults
of the CodeVersion and CodeVersioningPolicy constructors instead of
trying to apply its own defaults; that way they should always stay in
sync.

I also had to apply @version_no_warnings` to a bunch more functions in
our tests.