Ignore __module__ field on classes when hashing
When computing the hash a class, we now ignore the `__module__` field.
This field usually contains the name of the module the class was defined
in, but if that module was run directly from the command line then it
contains `__main__` instead. That means we can get different values
depending on whether we run a python module directly or load it from
somewhere else (like a Jupyter notebook), causing our smart caching to
detect a spurious change.