Ignore __name__ variable when hashing
This is basically the same problem as
ed262fbdc9bf9a2476f23b8d0cd5767fc8aeed56, but with the `__name__` global
variable instead of the `__module__` field of a class. I've tested this
on a complex Bionic flow and assisted versioning now seems to work
properly in both notebooks and on the command line.